fixes for compatibility with camlp5 8.03.00
For a while, instead of calling "camlp5 pr_xx.cmo pa_yy.cmo etc", you
can use "not-ocamlfind preprocess" and give it package-names.  This is
more upward-compatible, and hides the details of exactly how camlp5 is
invoked.

I've modified your Makefiles to do this, and left the original
invocations in comments.  Everything appears to build, and tests seem
to pass.  Though I'm not sure about the latter.